+++ src/entities/Ship.js
@@ -34,6 +34,8 @@ export class Ship extends Entity {
     this.miningLaserBeamLength = SHIP.MINING_LASER_RANGE;
     /** Seconds Space has been held (Precision warm-up) */
     this.mainEngineWarmup = 0;
+    /** Hangar hover altitude cue (1 = lifted; scales draw size) */
+    this.visualScale = 1;
   }
 
   getForward() {